legion/packager/rest/requirements.txt
click>=7.0
urllib3>=1.24.3
pyyaml>=3.1.2
pydantic>=0.29
docker
cloudstorage[amazon,google,microsoft]
google-cloud-storage>=1.18.0
botocore>=1.12.75
boto3>=1.9.75
jinja2>=2.7.3